move TestReceiverStatus to receiver_test.go
